pyproject.toml
[tool.poetry]
name = "hardy-calendar"
version = "0.1.0"
description = "Python script to synchronize hardy training plan with calendar"
authors = ["malyjasiak@gmail.com"]
readme = "README.md"
packages = [{include = "hardy_calendar", from = "src"}]

[tool.poetry.scripts]
hardy-calendar = "hardy_calendar.main:main"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.100.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"